backend/requirements.txt
aiosmtplib==2.0.2
alembic==1.11.2
amqp==5.1.1
annotated-types==0.5.0
anyio==3.7.1
async-timeout==4.0.3
asyncpg==0.28.0
bcrypt==4.0.1
billiard==4.1.0
black==23.7.0
blinker==1.6.2
celery==5.3.1
certifi==2023.7.22
cffi==1.15.1
click==8.1.6
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cryptography==41.0.3
dnspython==2.4.2
ecdsa==0.18.0
elastic-transport==8.4.0
elasticsearch==8.9.0
email-validator==2.0.0.post2
Faker==19.3.0
fastapi==0.101.1
fastapi-mail==1.4.1
greenlet==2.0.2
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.4
iniconfig==2.0.0
Jinja2==3.1.2
kombu==5.3.1
Mako==1.2.4
MarkupSafe==2.1.3
mypy-extensions==1.0.0
packaging==23.1
passlib==1.7.4
pathspec==0.11.2
platformdirs==3.10.0
pluggy==1.2.0
prompt-toolkit==3.0.39
psycopg2-binary==2.9.7
pyasn1==0.5.0
pycparser==2.21
pydantic==2.1.1
pydantic-settings==2.0.3
pydantic_core==2.4.0
pytest==7.4.0
pytest-asyncio==0.21.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jose==3.3.0
redis==4.6.0
rsa==4.9
ruff==0.0.287
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.19
starlette==0.27.0
typing_extensions==4.7.1
tzdata==2023.3
urllib3==1.26.16
uvicorn==0.23.2
vine==5.0.0
wcwidth==0.2.6
